+++ src/formatDate.ts
@@ -1,10 +1,10 @@
 import { resolveDateTimeFormatOptions } from "./internal/resolveDateTimeFormatOptions";
-import { createDateFormat } from "./internal/createDateFormat";
+import { createDateFormatIntlObj } from "./internal/createDateFormatIntlObj";
 import { DateIsoString } from "./types/dateIsoString";
 import { Locale } from "./types/locale";
 import { DateFormatPrecision } from "./types/dateFormatPrecision";
 import { convertToDate } from "./internal/convertToDate";
 
 export function formatDate(value: Date | DateIsoString, dateFormatType: DateFormatPrecision = "days", locale?: Locale): string {
-  return createDateFormat(locale, resolveDateTimeFormatOptions(dateFormatType)).format(convertToDate(value));
+  return createDateFormatIntlObj(locale, resolveDateTimeFormatOptions(dateFormatType)).format(convertToDate(value));
 }