+++ src/formatNumber.ts
@@ -1,45 +1,27 @@
-import { findIso4217Currency } from "./findIso4217Currency";
-import { createNumberFormat } from "./internal/createNumberFormat";
+import { createNumberFormat } from "./createNumberFormat";
 import { roundUsingBankersMethod } from "./roundUsingBankersMethod";
 import { Locale } from "./types/locale";
 import { NumberFormat } from "./types/numberFormat";
+import { NumberFormatIntl } from "./types/numberFormatIntl";
 
-export function formatNumber(value: number, options?: Partial<NumberFormat>, locale?: Locale): string {
+export function formatNumber(value: number, options?: Partial<NumberFormat>, locale?: Locale): string
+export function formatNumber(value: number, options?: NumberFormatIntl): string
+export function formatNumber(value: number, options?: Partial<NumberFormat> | NumberFormatIntl, locale?: Locale): string {
   if (options?.useBankersRounding && options?.maxPrecision !== undefined) {
     value = roundUsingBankersMethod(value, options.maxPrecision);
   }
+  
+  const formatOptions = (options as NumberFormatIntl)?.numberFormat !== undefined 
+    ? (options as NumberFormatIntl) 
+    : createNumberFormat(options, locale);
 
-  const formatter = createNumberFormat(locale, createFormatterOptions(options)),
-    formattedValue = formatter.format(value);
+  const formattedValue = formatOptions.numberFormat.format(value);
 
   if (options?.negativeZero === false && value <= 0) {
-    return formatter.format(-0) === formattedValue 
-      ? formatter.format(0) 
+    return formatOptions.numberFormat.format(-0) === formattedValue 
+      ? formatOptions.numberFormat.format(0) 
       : formattedValue;
   }
 
   return formattedValue;
-}
-
-function createFormatterOptions(opts?: Partial<NumberFormat>): Intl.NumberFormatOptions {
-  const minimumFractionDigits = opts?.minPrecision,
-    maximumFractionDigits = opts?.maxPrecision ?? 10,
-    style = opts?.currency ? "currency" : "decimal",
-    currency = opts?.currency !== undefined
-      ? typeof opts.currency === "string"
-        ? opts.currency
-        : findIso4217Currency(opts.currency)?.alpha3Code
-      : undefined;
-
-  return {
-    style: style,
-    currency: currency,
-    currencyDisplay: style === "currency" 
-      ? opts?.currencyDisplay ?? "symbol" 
-      : undefined,
-    currencySign: opts?.currencySign,
-    minimumFractionDigits: minimumFractionDigits,
-    maximumFractionDigits: maximumFractionDigits,
-    useGrouping:opts?.thousandsSeparator === false ? false : true
-  };
 }