+++ src/internal/getSeparator.ts
@@ -1,9 +1,9 @@
-import { createNumberFormat } from "./createNumberFormat";
+import { createNumberFormatIntlObj } from "./createNumberFormatIntlObj";
 import { Locale } from "../types/locale";
 
 /** @internal */
 export function getSeparator(value: number, separatorType: Intl.NumberFormatPartTypes, options?: Intl.NumberFormatOptions, locale?: Locale): string {
-  return createNumberFormat(locale, options)
+  return createNumberFormatIntlObj(locale, options)
     .formatToParts(value)
     .find(part => part.type === separatorType)?.value ?? "";
 }