+++ src/types/numberFormatIntl.ts
@@ -1,0 +1,5 @@
+import { NumberFormat } from "./numberFormat";
+
+export interface NumberFormatIntl extends Pick<Partial<NumberFormat>, "useBankersRounding" | "negativeZero" | "maxPrecision"> {
+  numberFormat: Intl.NumberFormat
+}